Version 0.0.42.001
Доделал наполнение окна вывода результатов

+++ design_of_mechanical_production/core/interfaces/i_workshop.py
@@ -13,7 +13,7 @@ class IWorkshop(Protocol):
     """
 
     name: str
-    production_volume: float
+    production_volume: Decimal
     mass_detail: Decimal
     process: 'IProcess'
     zones: Dict[str, 'IWorkshopZone']
